feat(utils): expose `toObject` helper;
- moved from internal-facing only
