+++ src/utils.test.ts
@@ -362,6 +362,42 @@ decode.run();
 
 // ---
 
+const toObject = suite('toObject');
+
+toObject('should be a function', () => {
+	assert.type(utils.toObject, 'function');
+});
+
+toObject('should match `Object.fromEntries` for simple values', () => {
+	let input = new URLSearchParams();
+	input.set('hello', 'world');
+	input.set('foo', 'bar');
+
+	assert.equal(
+		utils.toObject(input),
+		Object.fromEntries(input)
+	);
+});
+
+toObject('should keep multiple values per key', () => {
+	let input = new URLSearchParams();
+	input.append('hello', 'world');
+	input.append('hello', 'there');
+	input.set('foo', 'bar');
+
+	const output = utils.toObject(input);
+	assert.not.equal(output, Object.fromEntries(input));
+
+	assert.equal(output, {
+		foo: 'bar',
+		hello: ['world', 'there']
+	});
+});
+
+toObject.run();
+
+// ---
+
 const body = suite('body');
 
 const Request = (type?: string): any => ({
